% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graph.R
\name{BuildGraph}
\alias{BuildGraph}
\title{Build Graph from Seurat Object}
\usage{
BuildGraph(object, k = 20, reduction = "pca", dims = 1:30)
}
\arguments{
\item{object}{Seurat object}

\item{k}{Number of neighbors}

\item{reduction}{Reduction to use (default: "pca")}

\item{dims}{Dimensions to use}
}
\value{
A normalized adjacency matrix
}
\description{
Build Graph from Seurat Object
}
