% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mask.R
\name{CreatePathwayMask}
\alias{CreatePathwayMask}
\title{Create Pathway Mask from GMT}
\usage{
CreatePathwayMask(gset.idx.list, features, sign_k)
}
\arguments{
\item{gset.idx.list}{List of gene sets (e.g., read from GMT) where names are pathway/term names and values are gene vectors.}

\item{features}{Vector of gene names used in the model (must match order of input data).}

\item{sign_k}{Number of diffusion steps (to repeat the mask for diffused features).}
}
\value{
A binary matrix [Pathways x (Genes * (sign_k + 1))]
}
\description{
Helper to create a binary mask for scMSGNN from a GMT file or list of pathways.
}
