% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_tensor}
\alias{ensure_tensor}
\title{Convert to torch tensor if needed}
\usage{
ensure_tensor(x, device = "cpu", ...)
}
\arguments{
\item{x}{Matrix or tensor}

\item{device}{Device to use ("cpu" or "cuda")}

\item{...}{additional parameters}
}
\value{
torch_tensor
}
\description{
Convert to torch tensor if needed
}
