% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_torch_tensor}
\alias{is_torch_tensor}
\title{Check if object is a torch tensor}
\usage{
is_torch_tensor(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical
}
\description{
Check if object is a torch tensor
}
