% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_adjacency}
\alias{normalize_adjacency}
\title{Normalize adjacency matrix for GCN}
\usage{
normalize_adjacency(adj, add_self_loops = TRUE, device = "cpu")
}
\arguments{
\item{adj}{Adjacency matrix (tensor or matrix)}

\item{add_self_loops}{Whether to add self-loops}

\item{device}{Device to use}
}
\value{
Normalized adjacency tensor
}
\description{
Computes D^(-0.5) * A * D^(-0.5)
}
