% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graph.R
\name{NormalizeDegree}
\alias{NormalizeDegree}
\title{Normalize Adjacency Matrix}
\usage{
NormalizeDegree(adj)
}
\arguments{
\item{adj}{A sparse adjacency matrix (dgCMatrix)}
}
\value{
A normalized sparse matrix
}
\description{
Calculates the normalized Laplacian matrix: D^(-1/2) * (A + I) * D^(-1/2)
}
