% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plotCluster}
\alias{plotCluster}
\title{Plot Clustering Results}
\usage{
plotCluster(object, reduction = "msgnn", ...)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Reduction to use (default: "msgnn")}

\item{...}{Additional arguments passed to ggsc::sc_dim}
}
\description{
Wrapper around ggsc::sc_dim to plot clustering results on MSGNN embedding
}
