% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plotGraph}
\alias{plotGraph}
\title{Plot Graph Structure}
\usage{
plotGraph(object, layout = "fr")
}
\arguments{
\item{object}{Seurat object}

\item{layout}{Layout algorithm}
}
\value{
A ggplot object
}
\description{
Wrapper around ggtangle for network visualization
}
