% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plotScatter}
\alias{plotScatter}
\title{Plot Feature Scatter}
\usage{
plotScatter(object, features, reduction = "msgnn", ...)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Features to plot}

\item{reduction}{Reduction to use}

\item{...}{Additional arguments passed to ggsc::sc_feature}
}
\description{
Wrapper around ggsc::sc_feature
}
